server/requirements.txt
gunicorn==21.2.0
Werkzeug==2.3.7
Flask==2.3.3
Flask-Cors==4.0.0
firebase-admin==6.2.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0
pandas==2.0.3